pull Game out into app/game.js
